refactor: rename SidePanel to DetailPanel
